Initialize new instances of `Domain`, add `id` method
In cases where `Domain` is going to do something provider specific - add the `cloud` client on initialization to save adding logic to each method that requires a provider-specific action. In the case where it’s a generic thing across providers, such as `list` - we’ll still add that logic in.

Also add `id` method, which grabs the `cloudware_id` from a given `Domain` - rather than having this functionality within the provider