Initialize `Machine` with provider
In cases where `Machine` is going to do something provider specific - add the `cloud` client on initialization to save adding logic to each method that requires a provider-specific action. In the case where it’s a generic thing across providers, such as `list` - we’ll still add that logic in.