Move the `create machine` code into its on class
There is to much code in this file. The `machine create` code has been
moved to its own file.

The new `Command` class is the base class for all future commands. It
contains the basic structure required to run a command. The key part for
the inheritance is the `run` method. The `run!` is used by the cli.

Otherwise the `commands` file is to make requiring all the commands
easier.