Remove the `nil` error check on the config
This is a poor check as it will only be raised if the file is present
BUT empty. If it is missing, than a missing file error will be raised.

If the file contains any `YAML` (regardless if it is correct) it will
pass the check as well.

The best cause of action is to remove the check completely. At some
point it can then be replaced with a more robust validation.

https://github.com/alces-software/cloudware/pull/34#discussion_r187371426