Abstract away the updating of the Whirly status
This way the commands do not need to know how to update the status
themselves. Instead, all they need to do is run the provided proc.

The advantage of the implementation is any additional string processing
does not need to be duplicated across the code. Instead it can be done
within the single proc.

In this case, it ensures that the `status` is always bold