Define the base `create` method for models
The `create` method will be common to all models. It has been setup to
work with callbacks. This will allow additional validations to run on
create only.

The `run_create` method is designed to be inherited and contains the
code that will interact with the Providers API's.

It will only ever run if the `model` is valid. Otherwise it will skip
`run_create`. In all cases, it will return the model object.