Prevent the invalid IP error being repeated
Due to how the IP's are validated, the invalid `networkcdir` error would
be displayed twice. This is because it would be reran in the
`validate_networkcdir_contians_prisubnetcidr`.

The fix is a flag that adds the error by default. However it can be used
to turn off the add in latter checks