Extract the code for deploying onto aws
Both deploying a machine and a domain use similar cloud formation code.
In order to separate concerns however, the `domain` and `machine`
specific code will be located in different files.

However the shared deploy code has been put into a `DeployAWS` module.
This way it can be included into the relevant `domain` or `machine`
class.

This means `Models::Domain` is the base model class.
`Providers::DeployAWS` is responsible for interacting with the
`aws-sdk`. This then leaves `Providers::Domains::AWS` as an adapter
between the two.