Speed up the error if the domain already exists
Previously all the regions needed to be polled to check if the domain
name already exits. However the specification has been change to only
ensuring uniqueness within a region.

This means the domain doesn't need to be checked in advance if it
already exists. If it does, it will raise an error. This error is caught
by the `Prociders::Domains::AWS` class which sets a flag on
`Models::Domains` that results in the validation error.