Switch checking if a domain exists to the new model
Only domains that are created by `cloudware` should be deleted by
`cloudware`. As such all the domains are belong to the region are polled
and then checked for the input name.

If the domain does not exist, then it errors. This is to prevent
cloudware deleting domains it does not manage.

Currently the original domain object still does the deletion