Delete the stack through the new Provider object
This removes the need for the original `Domain` class in the command.

A bug needed to be fixed in where the method call on the
`cloud_formation` was not included. Also the `destroy` method needed to
be public.

The original `cloud` method needed to be added back into the
`Models::Domain` as it is needed in both `create` and `destroy`