Create a `Domains` class based on an array
It is a bit weird having the `by_region` method on the `Domain` class.
Instead a new `Domains` base class has been defined. This class can be
inherited from to build the domain list.

It will also contain handy filtering methods such as `find_by_name`

It has been setup that the inherited class need to override
`domains_array_by_region`. All this needs to do is return a simple array
of `Models::Domains` in the region. This is then transformed into the
correct `Domains` array object.