Move the domain existence check into the Model
During the destroy command, the domain must exist before it can be
destroyed. As this is always the case, it might as well be in the
`Domain` model.

As such, the check is ran `before_destroy`. A new `destroy!` method has
been added, that will raise and error if the domain does not exist. It
does not however validate the object fully before destroying it.