Add method for finding the regions
A method has been added onto the `aws` module for finding the regions.
Ironically, a region is required to find the regions as it determines
the end point for the call.

As there is no guarantee that the input nor default region belongs to
`aws`, the region must be hard coded. As such `us-west-1` has been
selected as it will unlikely ever go done.

The purpose of the `Base` module definition is to act as a reference.
Otherwise it isn't used