Pass in a ClusterIndex into the templates
There is an optional `ClusterIndex` parameter that can be used when
creating a domain. It is optional and unrelated to the template input.
If it is included, it is passed to the provider.

It is intended to be used with the `domainU` template, however this is
not enforced. If it is included with the domain0 template, the `aws`
create will fail.