Create a development mode for cloudware
The development mode sets up the full suite of gems and more
specifically requires `pry` so that it can be used latter.

Also `pry-byebug` has been included which adds break point functionality
to `pry`.

The `Whirly` does however need to be stopped in `dev` mode as it does
not work well from within `pry`