Add the `memoist` gem into the Model classes
The `memoist` gem is a direct extraction of `ActiveSupport::Memoize`
before it was deprecated. It provides the ability to memorize the return
values from a method without having to explicitly use instance
variables.

In this case, the `resource_group` has been memorized so it is only ever
created once. The same applies to the `deployment_model`