Add the code which will trigger the domain create
The code that triggers the domain to be created has been ported into the
new `Domain` class.

Note that the resource group name has been changed. Upon inspection of
the original `azure.rb` file, they where not the same. As such the
resource group name is now prepended with `domain`