Fix the deployment parameters for the domain
Various fixes required to get the domain creation working. Firstly the
deployment parameters must have a specific syntax in order to be
correctly encoded within the request (see code comment).

Secondly, the template file needs to be loaded as it is stored locally.
Otherwise azure gets the name of the template, not the template itself