Build the Domain models from Azure resources
Following the same design pattern as the `AWS` module, there is a
`Builder` class that allows `Domain` models to be created from Azure
resources.

Unlike `AWS`, azure can only download all the resource_groups globally.
This means the groups need to be filtered locally.

The `id` has also been moved to the base `Domain` model and
`attr_writer` as been added. This allows the `id` to be set from the
downloaded resource.