Dynamically pull the state for azure machines
The state can only be retrieved on an individual base on azure (I
think).

As such the code for retrieving the state has been defined in the model
itself. Azure does not have a way to retrieve the power status directly.
Instead it has be extracted from the logs. >:(

As this differs from AWS, the accessor method has been removed from the
base `Machine` class and defined on the `AWS::Machine` class directly