Extract the deployment parameters to the base class
The Azure and Aws templates have been standardised to accept the same
parameters. Note: Azure no longer uses the `*CIDR` parameters, these have
been changed to `*Cidr`.

This means the domain `deployment_parameters` are the same regardless of
platform and thus can be extracted to the base class.

This also insures the cluster index is passed into the azure templates
as well.