Extract the shared deployment parameters
The shared parameters for deploying machines as been extracted. This
way machines use similar parameters regardless of provider.

Additional parameters will be required depending on the provider and
thus the use of `super` is required.

NOTE: There is a bug where `deployment_parameters` has been used
interchangeable with `deploy_parameters`. This will be fixed in a future
commit.