Define an `all_region` method for retrieving domains
Due to the difference in how the providers sdk's work, the method for
pulling all the domains needs to be defined for each provider.

In the case of `AWS`, the domains need to be retrieved individually for
each region.

As this will be done on a provider basis, the `base.rb` file is no
longer required