Remove the results file during the destroy
The deployment doesn't exist any more and thus the results should be
removed

The `FileUtils.rm_f` method is used as it ignores errors (e.g. the file
doesn't exist anymore). NOTE: `FileUtils.rm` does not prompt before it
removes a file. The only difference is the error handling

NOTE: The was an error unpacking the `name`, it is actually in the first
position of the `argv` input