Create the Machine models from deployment results
Deployments may or may not create machines when they are deployed. The
catch with this is cloudware no longer guess how it should behave,
instead it must be explicitly told about the machines.

As such a new tag prefix has been added for notifying about nodes:
`cloudwareNodeID`

This is used to generate the list of machines that correspond to the
node