Abstract away setting the name from the tag
Eventually the `Machine` model will become a db record. When this
happens it should save the name NOT the tag. This is b/c it can build
the tag from the name.

However this requires external classes to convert the tag into a name.
As this is likely an error prone process, this has been abstracted away
by a custom attr writer.

This allows the name to be set from the tag w/o knowledge on how to do
this