Create the basic code structure for the Context
Eventually the context will be based of a file that stores the data.
However the key feature of the Context is to merge the results together

This commit implements the basic structure to preform this. ATM the
results are stored on the `Deployment` objects however this will be
changed to storing it on the context