Fake the filesystem during the tests
The file system needs to be faked within the tests to prevent the
development environment from affecting the test suite

This did require rejigging how the config path is set, to avoid a
chicken and egg situation when trying to retrieve it. This is b/c a
config file needs to be cloned into the fake filesystem for adminware to
work correctly.

Finally the faked filesystem is cleared after each spec which prevents
tests interacting with each other