Update the tag syntax for machines
Cloudware does not enforce a strict set of information to be stored
against a machine. Instead it has a new pattern it will recognise as
machine information.

The syntax for tagging an output against a machine is going to be:
`cloudwareNODE<name>TAG<tag-name>`

Currently a machine can be built off any tag, this will be updated