Use constants for each part of the tags
The tags now have two components, the node prefix and the tag flag. By
using constants for these components, multiple regex statements can stay
consistent

The `PROVIDER_ID_FLAG` has also been set as a constant. This is the
only enforced tag required for the `Machine` object to function
correctly. As such it is required outside of `Machine` in the spec