Implement the info domain command
Domains are going to work broadly in the same way as machines. As such
this code has been based off machines.

There is a new `Domain` model that can be created from a deployment like
a `Machine`. It also can have tags associated with it like a `Machine`.
The tag pattern is as follows:
`cloudwareDOMAIN<name>TAG<tag-name>`