Port the status code from the original cloudware
The code for finding the machine status has been ported from the
original version of cloudware

The only major change is how the resource group and machine names are
specified. The `Machine` class is initialized with the machine id.
However the azure sdk does not use the ID. But instead the ID encodes
the above name information which needs to be extracted first