Initialize the context with a region
When creating a new context, it can specify the region as an input. A
key word input is used as this is consistent with the other models.

It is only possible to specify the region on initialize as changing the
region should not be supported. This could lead to some weird bugs