Set the app name and version correctly
A new `version.rb` file has been created to set the version instead of
embedding it in the CLI file.

The program name for the CLI also needs to be set dynamically. This is
because each provider has a different binary name which can be inferred
from the `$PROGRAM_NAME` inbuilt variable