Remove the `build_context` method off the deployment
This has a bit of a code smell about it as it implies the `Deployment`
model should have some saving ability. However as this is going to be
hadnled at the command level, it is a bit odd.

Instead, the validator recreates the context when it comes to validating
if a previous deployment exists. This is probably the safest method as
it ensures a stale object isn't being used