Update the Data loader to work with IO
There is going to be a use case where file locking is going to be
required. The `Data` loader is more responsible for handling the content
and shouldn't handle all the weird and wonderful ways a file can be
opened.

Instead the `Data` class has been updated to work with `IO`s as the file
input. This will allow for the greatest flexibility without cluttering
`Data`