Remove the documentation about the log file
The log file is now setup implicitly and does require user
intervention. As such it shouldn't be documented in such a prominent
location.

There are still advanced configuration details in the example config

Also the default config location is created automatically. So it doesn't
need explicit handling in the spec