Extract the loading and saving of deployments
There are a few steps required for this to be possible. This is
cluttering the `update_deployments` method.

Instead the load and save components have been extracted so an `Updater`
class that set the attributes when required

NOTE: The file is being `rewind` so it is consistent with the truncation
in the save method. Otherwise load start at the current pointer where
save starts from the beginning