Create the `Context` with a cluster
The `Context` objects will now be initialized with a "cluster". This is
going to replace the "region" input as each cluster will be tied to a
region

NOTE: Note the `Deployment` object does not know which cluster it is
associated with. This means it has to assign a dummy cluster. This will
be fixed in the next commit