Make `FakeFS::Pathname` behave like a string
There is a bug where `FakeFS::Pathname` does not exactly behave like its
counter part: `Pathname`. `Pathname` can be happily used with `File` and
`FileUtils` interchangeable with strings.

However `FakeFs` is not mimicking this behaviour. Instead, the
`Pathname` is being patched so it can be implicitly converted into
a string.