Yield to allow for missing parameters to be input
The model will now merge what it has within `replacements` with the
values prompted for the missing parameters. It will then validate the
deployment again to make sure things are now okay before moving on. This
second validation also has the positive effect of clearing any previous
errors that might now be resolved.