Allow for parameter prompting for domains
This may not be utilised but for feature parity and the case where a
domain may need parameters replaced this could become useful. Also when
we eventually loosen the restrictions on domains the two methods (for
nodes and domains) may get merged so ensuring this works now is a bonus.