Replace Commander::Delegates with proxy calls
The major draw back to `Commander` is the argument vector must be known
in advance. Commander's `Runner` class is initialized on top of the
argument vector, and then the commands are applied to it.

The Delegates bastardise this by making them look like class methods.
But actually a `Runner` has been initialized around the ARGV vector.
This prevents multiple runners and thus multiple CLI's in the same
process.

Instead a set of "Proxy Delegates" have been devised. These will collect
the command definitions in advance but does not initialize a `Runner`.
This allows the set of commands to be called on any argument vector,
but playing back the proxy calls at a latter time.

It is not the most efficient way of running a command line, but is a
step forward to detangling Commander